Add zero_vars parameter to memoryUsage_DataForDerivatives
This means that one can choose to not rezero the parameters for
every call, although this is still done by default. This is
becausethe FDLR OrbitalBase is made of two TrialWavefunction
objects (and will itself be part of a TrialWavefunction object).
So when we call this function on the outer TrialWavefunction, we
may not want to zero the two innner TrialWavefunction objects too.